pkg/receipt: move under pkg/installation
Moving to somewhere more logically relevant.
