+++ pkg/installation/receipt/receipt.go
